Grammar and Diction correction
I dont have the raw so i cant compare the part i corrected, i just corrected it in a way so it is more enjoyable to read.
I corrected some sentences that dont have subject, clauses that are in reverse, and some sentence that have double verb. 

And paragraph at line 23, i really need the see the raw to know the right translation as it is quite a complex sentence.